mvvm restructured
