+++ lib/modules/user_repo/user_repo_impl.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:mvvm_sample/data/remote/api_service.dart';
-import 'package:mvvm_sample/models/users.dart';
-import 'package:mvvm_sample/repository/users_repo/user_repo.dart';
+import 'package:mvvm_sample/modules/users/users_model.dart';
+import 'package:mvvm_sample/modules/user_repo/user_repo.dart';
 
 class UserRepoImp extends UserRepo {
   ApiService apiService = ApiService();
